Extract functionality for generate into service class
